+++ attributes/webserver.rb
@@ -1,3 +1,3 @@
-default['railsapp']['webserver']['server_name'] = 'webapp.sample'
-default['railsapp']['webserver']['server_alias'] = %w(www.webapp.sample)
-default['railsapp']['webserver']['document_root'] = '/var/www/webapp/current/public'
+default['rubystack']['webserver']['server_name'] = 'webapp.sample'
+default['rubystack']['webserver']['server_alias'] = %w(www.webapp.sample)
+default['rubystack']['webserver']['document_root'] = '/var/www/webapp/current/public'