+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'railsapp'
+name             'rubystack'
 maintainer       'David Saenz Tagarro'
 maintainer_email 'david.saenz.tagarro@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures railsapp'
+description      'Installs/Configures rubystack'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
@@ -12,5 +12,5 @@ depends 'phusionpassenger'
 
 %w(debian ubuntu).each { |os| supports os }
 
-source_url 'https://github.com/dsaenztagarro/railsapp' if respond_to?(:source_url)
-issues_url 'https://github.com/dsaenztagarro/railsapp/issues' if respond_to?(:issues_url)
+source_url 'https://github.com/dsaenztagarro/rubystack' if respond_to?(:source_url)
+issues_url 'https://github.com/dsaenztagarro/rubystack/issues' if respond_to?(:issues_url)