+++ spec/database_spec.rb
@@ -1,15 +1,15 @@
 require 'spec_helper'
 
-describe 'railsapp::database' do
+describe 'rubystack::database' do
   let(:postgresql_package_version) { '9.4+170.pgdg14.04+1' }
   let(:postgresql_version) { '9.4' }
   let(:chef_run) do
     ChefSpec::SoloRunner.new(platform: 'ubuntu', version: '12.04') do |node|
       node.set['database']['postgresql']['package_version'] = postgresql_package_version
       node.set['database']['postgresql']['version'] = postgresql_version
-      node.set['railsapp']['db']['name'] = 'db_production'
-      node.set['railsapp']['db']['user']['name'] = 'admin'
-      node.set['railsapp']['db']['user']['password'] = 'admin1234'
+      node.set['rubystack']['db']['name'] = 'db_production'
+      node.set['rubystack']['db']['user']['name'] = 'admin'
+      node.set['rubystack']['db']['user']['password'] = 'admin1234'
     end.converge described_recipe
   end