+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-railsapp CHANGELOG
+rubystack CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the railsapp cookbook.
+This file is used to list changes made in each version of the rubystack cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of railsapp
+- [your_name] - Initial release of rubystack
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.